WocChat usability enhancements
if term.read() gives nil or "", the user canceled it so cancel config
Do a simple nick validation that the nickname isn't blank
Fix text resetting to the top of the screen
Listen to drag events for the scrollbar
Don't trim input lines
Don't attempt to send with an empty string.
Start to add in high precision support, it's disabled for now since it
causes many issues.